style(shopping card): rename Card components to ShoppingCard
