+++ src/pages/products/electronics.jsx
@@ -1,7 +1,7 @@
 import React from "react"
 import Link from "next/link"
 import StarRating from "@/components/Rating/StarRating"
-import Card from "@/components/Card/Card"
+import Card from "@/components/ShoppingCard/ShoppingCard"
 
 const buttonStyle = {
   backgroundColor: "#BC9B7F",
